[WIP] Pass all tests and doctest.
