+++ jageocoder/tree.py
@@ -1,11 +1,12 @@
-from collections import OrderedDict
+from abc import ABC, abstractmethod
+
 from logging import getLogger
 import os
 from pathlib import Path
 import re
 import site
 import sys
-from typing import Any, Dict, List, Optional, Set, Union
+from typing import Any, Dict, List, Optional, Set, Tuple, Union
 
 from deprecated import deprecated
 
@@ -88,30 +89,7 @@ def get_db_dir(mode: str = 'r') -> Optional[Path]:
     return None  # In case of read-only mode.
 
 
-class LRU(OrderedDict):
-    'Limit size, evicting the least recently looked-up key when full'
-
-    def __init__(self, maxsize=512, *args, **kwds):
-        self.maxsize = maxsize
-        super().__init__(*args, **kwds)
-
-    def __getitem__(self, key):
-        value = super().__getitem__(key)
-        self.move_to_end(key)
-        return value
-
-    def __setitem__(self, key, value):
-        if key in self:
-            self.move_to_end(key)
-
-        super().__setitem__(key, value)
-        if len(self) > self.maxsize:
-            oldest = next(iter(self))
-            logger.debug("Delete '{}'".format(oldest))
-            del self[oldest]
-
-
-class AddressTree(object):
+class AddressTree(ABC):
     """
     The address-tree structure.
 
@@ -139,72 +117,25 @@ class AddressTree(object):
         Converter object of character-variants.
     """
 
-    def __init__(self,
-                 db_dir: Optional[os.PathLike] = None,
-                 mode: str = 'a',
-                 debug: Optional[bool] = None):
+    def __init__(
+        self,
+        debug: Optional[bool] = None
+    ):
         """
         The initializer
 
         Parameters
         ----------
-        db_dir: os.PathLike, optional
-            The database directory.
-            If omitted, the directory returned by get_db_dir() is used.
-            'address.db' and 'address.trie' are stored under this directory.
-
-        mode: str, optional (default='a')
-            Specifies the mode for opening the database.
-
-            - In the case of 'a', if the database already exists,
-              use it. Otherwize create a new one.
-
-            - In the case of 'w', if the database already exists,
-              delete it first. Then create a new one.
-
-            - In the case of 'r', if the database already exists,
-              use it. Otherwise raise a JageocoderError exception.
-
         debug: bool, optional (default=False)
             Debugging flag. If set to True, write debugging messages.
             If omitted, refer 'JAGEOCODER_DEBUG' environment variable,
             or False if the environment variable is also undefined.
         """
-        # Set default values
-        self.mode = mode
-        if db_dir is None:
-            db_dir = get_db_dir(mode)
-        else:
-            db_dir = Path(db_dir).absolute()
-
-        if db_dir is None or not db_dir.is_dir():
-            msg = "Directory '{}' does not exist.".format(db_dir)
-            raise AddressTreeException(msg)
-
-        self.db_dir = db_dir
-        self.address_nodes: AddressNodeTable = AddressNodeTable(
-            db_dir=self.db_dir)
-        self.aza_masters: AzaMaster = AzaMaster(db_dir=self.db_dir)
-        self.trie_nodes: TrieNode = self.get_trie_nodes()
-        self.trie_path = db_dir / 'address.trie'
-
         # Options
         self.debug = debug or bool(os.environ.get('JAGEOCODER_DEBUG', False))
 
-        # Clear database when in write mode.
-        if self.mode == 'w':
-            self.address_nodes.delete()
-            if os.path.exists(self.trie_path):
-                os.remove(self.trie_path)
-
-        self.root = None
-        self.trie = AddressTrie(self.trie_path)
-        self.reverse_index = None
-
-        # Regular expression
-        self.re_float = re.compile(r'^\-?\d+\.?\d*$')
-        self.re_int = re.compile(r'^\-?\d+$')
-        self.re_address = re.compile(r'^(\d+);(.*)$')
+        # Itaiji converter
+        self.converter = Converter()
 
         # Set default settings
         self.config = {
@@ -231,35 +162,22 @@ class AddressTree(object):
                 self.config["auto_redirect"]),
         })
 
-        # Itaiji converter
-        self.converter = Converter()
-
-    def __not_in_readonly_mode(self) -> None:
-        """
-        Check if the dictionary is not opened in the read-only mode.
-
-        If the mode is read-only, AddressTreeException will be raised.
-        """
-        if self.mode == 'r':
-            raise AddressTreeException(
-                'This method is not available in read-only mode.')
-
-    def get_address_nodes(self) -> AddressNodeTable:
-        return self.address_nodes
-
     @property
+    @abstractmethod
     def datasets(self) -> Optional[Dict[int, Any]]:
         """
         Get list of datasets installed in the dictionary.
 
         Returns
         -------
-        List[Dataset]:
+        Dict[int, Any]:
             List of datasets.
         """
-        a = self.get_address_nodes()
-        b = a.datasets
-        return self.address_nodes.datasets.get_all()
+        pass
+
+    @abstractmethod
+    def get_azamasters(self) -> AzaMaster:
+        pass
 
     def get_root(self) -> AddressNode:
         """
@@ -290,6 +208,7 @@ class AddressTree(object):
 
         return root_node.note
 
+    @abstractmethod
     def get_node_by_id(self, node_id: int) -> AddressNode:
         """
         Get the full node information by its id.
@@ -303,10 +222,9 @@ class AddressTree(object):
         -------
         AddressNode
         """
-        node = self.address_nodes.get_record(node_id)
-        node.tree = self
-        return node
+        pass
 
+    @abstractmethod
     def search_nodes_by_codes(
             self,
             category: str,
@@ -325,17 +243,13 @@ class AddressTree(object):
         -------
         List[AddressNode]
         """
-        nodes = []
-        pattern = '{}:{}'.format(category, value)
-        nodes = self.address_nodes.search_records_on(
-            attr="note", value=pattern)  # exact match
+        pass
 
-        return nodes
-
+    @abstractmethod
     def search_ids_by_codes(
             self,
             category: str,
-            value: str) -> List[AddressNode]:
+            value: str) -> List[int]:
         """
         Search node ids by category and value.
 
@@ -350,15 +264,23 @@ class AddressTree(object):
         -------
         List[int]
         """
-        ids = []
-        pattern = '{}:{}'.format(category, value)
-        ids = self.address_nodes.search_ids_on(
-            attr="note", value=pattern)  # exact match
+        pass
 
-        return ids
+    @abstractmethod
+    def search_aza_records_by_codes(self, code: str) -> Any:
+        """
+        Search Address-base-registry's aza records.
 
-    def search_aza_records_by_codes(self, code: str):
-        return self.aza_masters.search_by_code(code)
+        Parameters
+        ----------
+        code: str
+            Machi-aza code in ABR.
+
+        Returns
+        -------
+        AzaRecord
+        """
+        pass
 
     @deprecated("Use 'node.get_fullname()' instead of this method.")
     def get_node_fullname(self, node: Union[AddressNode, int]) -> List[str]:
@@ -432,17 +354,6 @@ class AddressTree(object):
             for v in value:
                 if re.match(r'\d{2}', v) or re.match(r'\d{5}', v):
                     return
-
-                # Check if the value is a name of node in the database.
-                std = self.converter.standardize(v)
-                candidates = self.trie.common_prefixes(std)
-                if std in candidates:
-                    trie_node_id = candidates[std]
-                    for node_id in self.trie_nodes.get_record(
-                            pos=trie_node_id).nodes:
-                        node = self.address_nodes.get_record(pos=node_id)
-                        if node.name == v:
-                            return
 
                 msg = "'{}' is not a valid value for {}.".format(v, key)
                 raise RuntimeError(msg)
@@ -560,12 +471,13 @@ class AddressTree(object):
 
         >>> import jageocoder
         >>> jageocoder.init()
+        >>> jageocoder.get_module_tree().set_config(aza_skip=True)
         >>> jageocoder.get_module_tree().get_config('aza_skip')
-        'off'
+        True
         >>> jageocoder.get_module_tree().get_config(['best_only', 'target_area'])
         {'best_only': True, 'target_area': []}
         >>> jageocoder.get_module_tree().get_config()
-        {'debug': False, 'aza_skip': 'off', 'best_only': True, 'target_area': [], 'require_coordinates': False}
+        {'debug': False, 'aza_skip': True, 'best_only': True, 'target_area': [], 'require_coordinates': True, 'auto_redirect': True}
         """  # noqa: E501
         if keys is None:
             return self.config
@@ -584,255 +496,8 @@ class AddressTree(object):
             raise RuntimeError(
                 "The config key '{}' does not exist.".format(key))
 
-        return self.config[key]
-
-    def get_trie_nodes(self) -> TrieNode:
-        """
-        Get the TRIE node table.
-
-        Notes
-        -----
-        - Todo: If the trie index is not created, create.
-        """
-        return TrieNode(db_dir=self.db_dir)
-
-    def search_by_tree(self, address_names: List[str]) -> AddressNode:
-        """
-        Get the corresponding node from the list of address element names,
-        recursively search for child nodes using the tree.
-
-        For example, ['東京都','新宿区','西新宿','二丁目'] will search
-        the '東京都' node under the root node, search the '新宿区' node
-        from the children of the '東京都' node. Repeat this process and
-        return the '二丁目' node which is a child of '西新宿' node.
-
-        Parameters
-        ----------
-        address_names : list of str
-            A list of address element names to be searched.
-
-        Return
-        ------
-        AddressNode:
-            The node matched last.
-        """
-        cur_node = self.get_root()
-        for name in address_names:
-            name_index = self.converter.standardize(name)
-            node = cur_node.get_child(name_index)
-            if not node:
-                break
-            else:
-                cur_node = node
-
-        return cur_node
-
-    def search_by_trie(
-        self,
-        query: str,
-        processed_nodes: Optional[Set[int]] = None
-    ) -> dict:
-        """
-        Get the list of corresponding nodes using the TRIE index.
-        Returns a list of address element nodes that match
-        the query string in the longest part from the beginning.
-
-        For example, '中央区中央1丁目' will return the nodes
-        corresponding to '千葉県千葉市中央区中央一丁目' and
-        '神奈川県相模原市中央区中央一丁目'.
-
-        Parameters
-        ----------
-        query : str
-            An address notation to be searched.
-        processed_nodes: Set of the AddressNode's id, optional
-            List of node's id that have already been processed.
-
-        Return
-        ------
-        A dict object whose key is a node id
-        and whose value is a list of node and substrings
-        that match the query.
-        """
-        logger.debug((
-            "Called with query:'{}', processed_nodes:{}".format(
-                query,
-                processed_nodes
-            )))
-        index = self.converter.standardize(
-            query, keep_numbers=True)
-        index_for_trie = self.converter.standardize(query)
-        candidates = self.trie.common_prefixes(index_for_trie)
-        results = {}
-        max_len = 0
-        min_part = None
-        best_only = self.get_config('best_only')
-        target_area = self.get_config('target_area')
-
-        keys = sorted(candidates.keys(),
-                      key=len, reverse=True)
-
-        logger.debug("Trie: {}".format(','.join(keys)))
-
-        min_key = ''
-        _processed_nodes: Set[int] = processed_nodes or set()
-        resolved_node_ids: Set[int] = set()
-
-        for k in keys:
-            if len(k) < len(min_key):
-                logger.debug("Key '{}' is shorter than '{}'".format(
-                    k, min_key))
-                continue
-
-            trie_id = candidates[k]
-            logger.debug("Trie_id of key '{}' = {}".format(
-                k, trie_id))
-            trie_node = self.trie_nodes.get_record(pos=trie_id)
-            offset = self.converter.match_len(index, k)
-            key = index[0:offset]
-            rest_index = index[offset:]
-            for node_id in trie_node.nodes:
-                node = self.get_node_by_id(node_id=node_id)
-
-                if not node.has_valid_coordinate_values() \
-                        and self.get_config('require_coordinates'):
-                    node = node.add_dummy_coordinates()
-
-                if min_key == '' and node.level <= AddressLevel.WARD:
-                    # To make the process quicker, once a node higher
-                    # than the city level is found, addresses shorter
-                    # than the node are not searched after this.
-                    logger.debug((
-                        "A node with ward or higher levels found. "
-                        "Set min_key to '{}'").format(k))
-                    min_key = k
-
-                if node_id in _processed_nodes:
-                    logger.debug("Node {}({}) already processed.".format(
-                        node.name, node.id))
-                    continue
-
-                if len(target_area) > 0:
-                    # Check if the node is inside the specified area
-                    for area in target_area:
-                        inside = node.is_inside(area)
-                        if inside in (1, -1):
-                            break
-
-                    if inside == 0:
-                        msg = "Node {}({}) is not in the target area."
-                        logger.debug(msg.format(node.name, node.id))
-                        continue
-
-                logger.debug((
-                    "Search for the node with the longest match "
-                    "to the remaining '{}' recursively, "
-                    "starting with node '{}'(id:{})."
-                ).format(
-                    rest_index, node.name, node.id))
-                results_by_node = node.search_recursive(
-                    tree=self,
-                    index=rest_index,
-                    processed_nodes=_processed_nodes)
-                _processed_nodes.add(node_id)
-                logger.debug('{}({}) marked as processed'.format(
-                    node.name, node.id))
-
-                if len(results_by_node[0].matched) == 0 and \
-                        node.level == AddressLevel.CITY and \
-                        not rest_index.startswith(AddressNode.NONAME):
-
-                    logger.debug(
-                        "Search for NONAME Oaza of '{}'({})".format(
-                            node.name, node.id
-                        ))
-
-                    aza_skip = self.get_config('aza_skip')
-                    for result in results.values():
-                        if result[1].startswith(key) and result[1] > key:
-                            self.set_config(aza_skip=False)
-                            logger.debug(
-                                "Since one or more candidates are found, "
-                                "no omission of Aza will be checked."
-                            )
-                            break
-
-                    noname_child = node.get_table().get_record(pos=node.id + 1)
-                    if noname_child.name == AddressNode.NONAME and \
-                            noname_child.id not in _processed_nodes:
-                        _processed_nodes.add(noname_child.id)
-                        # Search under NONAME oaza node.
-                        for result in noname_child.search_recursive(
-                            tree=self,
-                            index=rest_index,
-                            processed_nodes=_processed_nodes
-                        ):
-                            if len(result.matched) > 0:
-                                results_by_node.append(result)
-                                logger.debug(
-                                    "Found '{}'({}) in NONAME Oaza.".format(
-                                        result.get_node().name,
-                                        result.get_node().id
-                                    ))
-
-                    self.set_config(aza_skip=aza_skip)
-
-                for cand in results_by_node:
-                    node = cand.get_node()
-                    if len(target_area) > 0:
-                        for area in target_area:
-                            inside = node.is_inside(area)
-                            if inside == 1:
-                                break
-
-                        if inside != 1:
-                            msg = "Node {}({}) is not in the target area."
-                            logger.debug(msg.format(node.name, node.id))
-                            continue
-
-                    if self.get_config("require_coordinates") \
-                            and not node.has_valid_coordinate_values():
-                        logger.debug("Node {}({}) has no coordinates.".format(
-                            node.name, node.id
-                        ))
-                        continue
-
-                    _len = offset + cand.nchars
-                    _part = offset + len(cand.matched)
-                    msg = "candidate: {} ({})"
-                    logger.debug(msg.format(key + cand.matched, _len))
-                    if best_only:
-                        if _len > max_len:
-                            results = {
-                                node.id: [cand.node, key + cand.matched]
-                            }
-                            max_len = _len
-                            min_part = _part
-
-                        elif _len == max_len and node.id not in results \
-                                and (min_part is None or _part <= min_part):
-                            results[node.id] = [
-                                cand.node, key + cand.matched]
-                            min_part = _part
-
-                    else:
-                        if node.id in resolved_node_ids:
-                            continue
-
-                        cur = node.get_parent()
-                        while cur is not None:
-                            resolved_node_ids.add(cur.id)
-                            cur = cur.parent
-
-                        results[node.id] = [
-                            node, key + cand.get_matched_string()]
-                        max_len = max(_len, max_len)
-                        if min_part is None:
-                            min_part = _part
-                        else:
-                            min_part = min(min_part, _part)
-
-        return results
+        v = self.config[key]
+        return v
 
     @deprecated(reason="Use 'get_node_by_id'.", version="2.1.7")
     def get_address_node(self, id: int) -> AddressNode:
@@ -849,8 +514,7 @@ class AddressTree(object):
         AddressNode
             Node with the specified ID.
         """
-        node = self.address_nodes.get_record(pos=id)
-        node.tree = self
+        node = self.get_node_by_id(node_id=id)
         return node
 
     @deprecated(('Renamed to `searchNode()` because it was confusing'
@@ -858,6 +522,7 @@ class AddressTree(object):
     def search(self, query: str, **kwargs) -> list:
         return self.searchNode(query, **kwargs)
 
+    @abstractmethod
     def searchNode(self, query: str) -> List[Result]:
         """
         Searches for address nodes corresponding to an address notation
@@ -879,103 +544,11 @@ class AddressTree(object):
         >>> jageocoder.init()
         >>> tree = jageocoder.get_module_tree()
         >>> tree.searchNode('多摩市落合1-15-2')
-        [[[11460207:東京都(139.69178,35.68963)1(lasdec:130001/jisx0401:13)]>[12063502:多摩市(139.446366,35.636959)3(jisx0402:13224)]>[12065383:落合(139.427097,35.624877)5(None)]>[12065384:一丁目(139.427097,35.624877)6(None)]>[12065390:15番地(139.428969,35.625779)7(None)], '多摩市落合1-15-']]
+        [{"node": {"id": ..., "name": "2", "name_index": "2.", "x": 139.4..., "y": 35.6..., "level": 8, "priority": 9, "note": "", "parent_id": ..., "sibling_id": ...}, "matched": "多摩市落合1-15-2"}]
         """  # noqa: E501
-        results = self.search_by_trie(query=query)
-        values = sorted(results.values(), reverse=True,
-                        key=lambda v: len(v[1]))
+        pass
 
-        matched_substring = {}
-        results = []
-        for v in values:
-            if v[1] in matched_substring:
-                matched = matched_substring[v[1]]
-            else:
-                matched = self._get_matched_substring(query, v)
-                matched_substring[v[1]] = matched
-
-            results.append(Result(v[0], matched))
-
-        # Sort the result list in descending order of the length of the match
-        # and ascending order of the node priority.
-        results.sort(
-            key=lambda r: len(r.matched) * -100 + r.node.priority)
-
-        return results
-
-    def _get_matched_substring(
-            self, query: str,
-            retrieved: list) -> str:
-        """
-        From the matched standardized substring,
-        recover the corresponding substring of
-        the original search string.
-
-        Parameters
-        ----------
-        query : str
-            The original search string.
-        matchd : str
-            The matched standardized substring.
-
-        Return
-        ------
-        str:
-            The recovered substring.
-        """
-        recovered = None
-        node, matched = retrieved
-        l_result = len(matched)
-        pos = l_result if l_result <= len(query) else len(query)
-        pos_history = [pos]
-
-        while True:
-            substr = query[0:pos]
-            standardized = self.converter.standardize(
-                substr, keep_numbers=True)
-            l_standardized = len(standardized)
-
-            if l_standardized == l_result:
-                recovered = substr
-                break
-
-            if l_standardized <= l_result:
-                pos += 1
-            else:
-                pos -= 1
-
-            if pos < 0 or pos > len(query):
-                break
-
-            if pos in pos_history:
-                message = "Can't de-standardize matched {} in {}".format(
-                    matched, query)
-                raise AddressTreeException(message)
-
-        if pos < len(query) and node.name != '':
-            if query[pos] == node.name[-1] and \
-                    len(self.converter.standardize(
-                        query[0:pos+1])) == l_result:
-                # When the last letter of a node name is omitted
-                # by normalization, and if the query string contains
-                # that letter, it is determined to have matched
-                # up to that letter.
-                # Ex. "兵庫県宍粟市山崎町上ノ１５０２" will match "上ノ".
-                recovered = query[0:pos+1]
-            elif query[-2:] in ('通り', '通リ'):
-                # '通' can be expressed as '通り'
-                recovered = query[0:pos+1]
-
-        return recovered
-
-    def create_note_index_table(self) -> None:
-        """
-        Collect notes from all address elements and create
-        search table with index.
-        """
-        self.__not_in_readonly_mode()
-        self.address_nodes.create_indexes()
-
+    @abstractmethod
     def reverse(
         self,
         x: float,
@@ -1007,11 +580,7 @@ class AddressTree(object):
         - Each element is a dict type with the following structure:
             {"candidate":AddressNode, "dist":float}
         """
-        if self.reverse_index is None:
-            from jageocoder.rtree import Index
-            self.reverse_index = Index(tree=self)
-
-        return self.reverse_index.nearest(x=x, y=y, level=level, as_dict=as_dict)
+        pass
 
     @classmethod
     def _clean_numerical_string(cls, code: str) -> str:
@@ -1056,12 +625,14 @@ class AddressTree(object):
                 return []
 
             citynode = citynode[0]
-            candidates = self.search_ids_by_codes(
+            candidates = self.search_nodes_by_codes(
                 category="aza_id",
                 value=id[-7:])
-            nodes = [self.address_nodes.get_record(x)
-                     for x in candidates
-                     if x >= citynode.id and x < citynode.sibling_id]
+            nodes = []
+            for node in candidates:
+                if node.id >= citynode.id and node.id < citynode.sibling_id:
+                    nodes.append(node)
+
         elif len(id) == 13:
             # lasdec(6digits) + aza_id(7digits)
             citynode = self.search_by_citycode(code=id[0:6])
@@ -1069,12 +640,14 @@ class AddressTree(object):
                 return []
 
             citynode = citynode[0]
-            candidates = self.search_ids_by_codes(
+            candidates = self.search_nodes_by_codes(
                 category="aza_id",
                 value=id[-7:])
-            nodes = [self.address_nodes.get_record(x)
-                     for x in candidates
-                     if x >= citynode.id and x < citynode.sibling_id]
+            nodes = []
+            for node in candidates:
+                if node.id >= citynode.id and node.id < citynode.sibling_id:
+                    nodes.append(node)
+
         else:
             nodes = self.search_nodes_by_codes(
                 category="aza_id",
